requirements.txt
absl-py==1.0.0
altair==4.1.0
anyio==3.4.0
argon2-cffi @ file:///C:/ci/argon2-cffi_1613037869401/work
asgiref==3.4.1
astor==0.8.1
astunparse==1.6.3
async-generator @ file:///home/ktietz/src/ci/async_generator_1611927993394/work
attrs @ file:///tmp/build/80754af9/attrs_1620827162558/work
backcall==0.2.0
base58==2.1.1
bleach @ file:///tmp/build/80754af9/bleach_1628110601003/work
blinker==1.4
cachetools==4.2.4
certifi==2021.10.8
cffi @ file:///C:/ci/cffi_1636523946399/work
charset-normalizer==2.0.8
click==7.1.2
colorama==0.4.4
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.0
defusedxml @ file:///tmp/build/80754af9/defusedxml_1615228127516/work
entrypoints==0.3
fastapi==0.70.0
flatbuffers==2.0
fonttools==4.28.2
gast==0.4.0
gitdb==4.0.9
GitPython==3.1.24
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.42.0
h11==0.12.0
h5py==3.6.0
httptools==0.2.0
idna==3.3
importlib-metadata @ file:///C:/ci/importlib-metadata_1631916826748/work
ipykernel==6.5.1
ipython==7.30.0
ipython-genutils @ file:///tmp/build/80754af9/ipython_genutils_1606773439826/work
ipywidgets @ file:///tmp/build/80754af9/ipywidgets_1634143127070/work
jedi==0.18.1
Jinja2 @ file:///tmp/build/80754af9/jinja2_1635780242639/work
joblib==1.1.0
jsonschema @ file:///Users/ktietz/demo/mc3/conda-bld/jsonschema_1630511932244/work
jupyter @ file:///C:/ci/jupyter_1607685287094/work
jupyter-client==7.1.0
jupyter-console @ file:///tmp/build/80754af9/jupyter_console_1616615302928/work
jupyter-core==4.9.1
jupyterlab-pygments @ file:///tmp/build/80754af9/jupyterlab_pygments_1601490720602/work
jupyterlab-widgets @ file:///tmp/build/80754af9/jupyterlab_widgets_1609884341231/work
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
libclang==12.0.0
Markdown==3.3.6
MarkupSafe @ file:///C:/ci/markupsafe_1621528502553/work
matplotlib==3.5.0
matplotlib-inline==0.1.3
mistune @ file:///C:/ci/mistune_1607359457024/work
nbclient @ file:///tmp/build/80754af9/nbclient_1614364831625/work
nbconvert @ file:///C:/ci/nbconvert_1624479160025/work
nbformat @ file:///tmp/build/80754af9/nbformat_1617383369282/work
nest-asyncio==1.5.1
notebook @ file:///C:/ci/notebook_1637161941035/work
numpy==1.21.4
oauthlib==3.1.1
opt-einsum==3.3.0
packaging @ file:///tmp/build/80754af9/packaging_1637314298585/work
pandas==1.3.5
pandocfilters @ file:///C:/ci/pandocfilters_1605114832805/work
parso==0.8.2
pickleshare==0.7.5
Pillow==8.4.0
prometheus-client @ file:///tmp/build/80754af9/prometheus_client_1637050397234/work
prompt-toolkit==3.0.23
protobuf==3.19.1
pyarrow==6.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic==1.8.2
pydeck==0.7.1
Pygments==2.10.0
Pympler==0.9
pyparsing @ file:///tmp/build/80754af9/pyparsing_1635766073266/work
pyrsistent @ file:///C:/ci/pyrsistent_1636093225342/work
python-dateutil==2.8.2
python-dotenv==0.19.2
python-multipart==0.0.5
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
pywin32==302
pywinpty @ file:///C:/ci/pywinpty_1607419945780/work
PyYAML==6.0
pyzmq==22.3.0
qtconsole @ file:///tmp/build/80754af9/qtconsole_1632739723211/work
QtPy @ file:///tmp/build/80754af9/qtpy_1629397026935/work
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
scikit-learn==1.0.1
scipy==1.7.3
Send2Trash @ file:///tmp/build/80754af9/send2trash_1632406701022/work
setuptools-scm==6.3.2
sip==4.19.13
six==1.16.0
sklearn==0.0
smmap==5.0.0
sniffio==1.2.0
starlette==0.16.0
streamlit==1.2.0
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow-estimator==2.7.0
tensorflow-gpu==2.7.0
tensorflow-io-gcs-filesystem==0.22.0
termcolor==1.1.0
terminado==0.9.4
testpath @ file:///tmp/build/80754af9/testpath_1624638946665/work
threadpoolctl==3.0.0
toml==0.10.2
tomli==1.2.2
toolz==0.11.2
tornado==6.1
traitlets==5.1.1
typing_extensions==4.0.0
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.7
uvicorn==0.15.0
validators==0.18.2
watchdog==2.1.6
watchgod==0.7
wcwidth==0.2.5
webencodings==0.5.1
websockets==10.1
Werkzeug==2.0.2
widgetsnbextension @ file:///C:/ci/widgetsnbextension_1607531582688/work
wincertstore==0.2
wrapt==1.13.3
zipp @ file:///tmp/build/80754af9/zipp_1633618647012/work
